# docs/openapi/openapi.v0.2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SKDY RAG Server API",
    "description": "HTTP API for the SKDY RAG Server. Mounted under `/api/v1`.\n\n**Contract**: the OpenAPI schema is the single source of truth for the Web UI; client types are generated from it.\n\n**Error envelope**: every 4xx/5xx returns `{\"error\": {\"code\": \"...\", \"message\": \"...\", \"request_id\": \"...\", \"details\": {}}}`.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/mcp-keys": {
      "get": {
        "tags": [
          "mcp-keys"
        ],
        "summary": "List MCP API keys",
        "operationId": "list_mcp_keys_api_v1_mcp_keys_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPKeyListResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "mcp-keys"
        ],
        "summary": "Create an MCP API key",
        "operationId": "create_mcp_key_api_v1_mcp_keys_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MCPKeyCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPKeySecretResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mcp-keys/{name}": {
      "patch": {
        "tags": [
          "mcp-keys"
        ],
        "summary": "Rename an MCP API key",
        "operationId": "rename_mcp_key_api_v1_mcp_keys__name__patch",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MCPKeyRenameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPKeyMetadata"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "mcp-keys"
        ],
        "summary": "Delete an MCP API key",
        "operationId": "delete_mcp_key_api_v1_mcp_keys__name__delete",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mcp-keys/{name}/collections": {
      "put": {
        "tags": [
          "mcp-keys"
        ],
        "summary": "Update an MCP API key collection whitelist",
        "operationId": "update_mcp_key_collections_api_v1_mcp_keys__name__collections_put",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MCPKeyCollectionsUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPKeyMetadata"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mcp-keys/{name}/rotate": {
      "post": {
        "tags": [
          "mcp-keys"
        ],
        "summary": "Rotate an MCP API key",
        "operationId": "rotate_mcp_key_api_v1_mcp_keys__name__rotate_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPKeySecretResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/mcp-keys/{name}/revoke": {
      "post": {
        "tags": [
          "mcp-keys"
        ],
        "summary": "Revoke an MCP API key",
        "operationId": "revoke_mcp_key_api_v1_mcp_keys__name__revoke_post",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MCPKeyMetadata"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MCPKeyCollectionsUpdateRequest": {
        "properties": {
          "allowed_collections": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Allowed Collections"
          }
        },
        "type": "object",
        "required": [
          "allowed_collections"
        ],
        "title": "MCPKeyCollectionsUpdateRequest"
      },
      "MCPKeyCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Name"
          },
          "allowed_collections": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Allowed Collections"
          }
        },
        "type": "object",
        "required": [
          "name",
          "allowed_collections"
        ],
        "title": "MCPKeyCreateRequest"
      },
      "MCPKeyListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MCPKeyMetadata"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "MCPKeyListResponse"
      },
      "MCPKeyMetadata": {
        "properties": {
          "key_id": {
            "type": "string",
            "title": "Key Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "allowed_collections": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Collections"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          }
        },
        "type": "object",
        "required": [
          "key_id",
          "name",
          "allowed_collections",
          "enabled",
          "created_at",
          "revoked_at",
          "last_used_at"
        ],
        "title": "MCPKeyMetadata"
      },
      "MCPKeyRenameRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "MCPKeyRenameRequest"
      },
      "MCPKeySecretResponse": {
        "properties": {
          "key_id": {
            "type": "string",
            "title": "Key Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "allowed_collections": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed Collections"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "api_key": {
            "type": "string",
            "format": "password",
            "title": "Api Key",
            "description": "The MCP Client Key. Returned exactly once at create/rotate time and never readable again; send-only for the connection test.",
            "writeOnly": true
          }
        },
        "type": "object",
        "required": [
          "key_id",
          "name",
          "allowed_collections",
          "enabled",
          "created_at",
          "revoked_at",
          "last_used_at",
          "api_key"
        ],
        "title": "MCPKeySecretResponse",
        "description": "Returned only by create/rotate; ``api_key`` is never listable."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}