# src/web_api/routers/mcp_keys.py
"""Trusted-admin HTTP API for MCP API-key management.

Authentication for this management surface is intentionally deferred by the
product roadmap. Deploy it only behind the same trusted admin boundary as the
Web UI; no endpoint returns an existing secret or digest.
"""

from __future__ import annotations

from fastapi import APIRouter, Response, status

from src.core.settings import load_settings
from src.mcp_server.auth import ApiKeyService
from src.application.services.audit_store import AuditStore
from pathlib import Path
from src.mcp_server.auth.models import (
    DuplicateKeyNameError, InvalidCollectionWhitelistError, KeyNotFoundError,
)
from src.web_api.errors import BadRequestError, ConflictError, NotFoundError
from src.web_api.schemas.mcp_keys import (
    MCPKeyCollectionsUpdateRequest, MCPKeyCreateRequest, MCPKeyListResponse,
    MCPKeyMetadata, MCPKeyRenameRequest, MCPKeySecretResponse,
)

router = APIRouter(prefix="/mcp-keys", tags=["mcp-keys"])


def _service() -> ApiKeyService:
    settings = load_settings("./config/settings.yaml")
    return ApiKeyService(
        db_path=settings.mcp_access.database_path,
        audit_store=AuditStore(Path("./data/db/audit.db")),
    )


def _metadata(value) -> MCPKeyMetadata:
    return MCPKeyMetadata(
        key_id=value.key_id, name=value.name,
        allowed_collections=sorted(value.allowed_collections), enabled=value.enabled,
        created_at=value.created_at, revoked_at=value.revoked_at,
        last_used_at=value.last_used_at,
    )


@router.get("", response_model=MCPKeyListResponse, summary="List MCP API keys")
async def list_mcp_keys() -> MCPKeyListResponse:
    return MCPKeyListResponse(items=[_metadata(item) for item in _service().list_keys()])


@router.post("", response_model=MCPKeySecretResponse, status_code=status.HTTP_201_CREATED,
             summary="Create an MCP API key")
async def create_mcp_key(body: MCPKeyCreateRequest) -> MCPKeySecretResponse:
    try:
        raw_key, item = _service().create_key(
            name=body.name, allowed_collections=set(body.allowed_collections),
        )
    except DuplicateKeyNameError as exc:
        raise ConflictError(str(exc)) from exc
    except (InvalidCollectionWhitelistError, ValueError) as exc:
        raise BadRequestError(str(exc)) from exc
    return MCPKeySecretResponse(**_metadata(item).model_dump(), api_key=raw_key)


@router.patch("/{name}", response_model=MCPKeyMetadata, summary="Rename an MCP API key")
async def rename_mcp_key(name: str, body: MCPKeyRenameRequest) -> MCPKeyMetadata:
    try:
        return _metadata(_service().rename_key(name=name, new_name=body.name))
    except KeyNotFoundError as exc:
        raise NotFoundError(str(exc)) from exc
    except DuplicateKeyNameError as exc:
        raise ConflictError(str(exc)) from exc
    except ValueError as exc:
        raise BadRequestError(str(exc)) from exc


@router.put("/{name}/collections", response_model=MCPKeyMetadata,
            summary="Update an MCP API key collection whitelist")
async def update_mcp_key_collections(
    name: str, body: MCPKeyCollectionsUpdateRequest,
) -> MCPKeyMetadata:
    try:
        return _metadata(_service().update_key_collections(
            name=name, allowed_collections=set(body.allowed_collections),
        ))
    except KeyNotFoundError as exc:
        raise NotFoundError(str(exc)) from exc
    except (InvalidCollectionWhitelistError, ValueError) as exc:
        raise BadRequestError(str(exc)) from exc

@router.post("/{name}/rotate", response_model=MCPKeySecretResponse,
             summary="Rotate an MCP API key")
async def rotate_mcp_key(name: str) -> MCPKeySecretResponse:
    try:
        raw_key, item = _service().rotate_key(name=name)
    except KeyNotFoundError as exc:
        raise NotFoundError(str(exc)) from exc
    return MCPKeySecretResponse(**_metadata(item).model_dump(), api_key=raw_key)


@router.post("/{name}/revoke", response_model=MCPKeyMetadata,
             summary="Revoke an MCP API key")
async def revoke_mcp_key(name: str) -> MCPKeyMetadata:
    try:
        return _metadata(_service().revoke_key(name=name))
    except KeyNotFoundError as exc:
        raise NotFoundError(str(exc)) from exc


@router.delete("/{name}", status_code=status.HTTP_204_NO_CONTENT, summary="Delete an MCP API key")
async def delete_mcp_key(name: str) -> Response:
    try:
        _service().delete_key(name=name)
    except KeyNotFoundError as exc:
        raise NotFoundError(str(exc)) from exc
    return Response(status_code=status.HTTP_204_NO_CONTENT)


__all__ = ["router"]